moving additional functionality to uploader_utils to be shared by profile uploader
